# doc/swagger/shortener.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shortener.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Shortener"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/url": {
      "post": {
        "operationId": "Shortener_PostUrl",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shortenerShortUrl"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/shortenerLongUrl"
            }
          }
        ],
        "tags": [
          "Shortener"
        ]
      }
    },
    "/v1/url/{shortUrl}": {
      "get": {
        "operationId": "Shortener_GetUrl",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/shortenerLongUrl"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shortUrl",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Shortener"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "shortenerLongUrl": {
      "type": "object",
      "properties": {
        "longUrl": {
          "type": "string"
        }
      }
    },
    "shortenerShortUrl": {
      "type": "object",
      "properties": {
        "shortUrl": {
          "type": "string"
        }
      }
    }
  }
}